pipeline/etl/requirements.txt
requests
ijson
zstandard
pytest
pytest-mock
psutils